requirements.txt
Django==5.0.6
python-dotenv==1.0.1
djangorestframework==3.15.2
django-rosetta==0.10.0
modeltranslation==0.25
phonenumbers==8.13.40
django-phonenumber-field==8.0.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9